requirements.txt
chromadb==1.5.1
joblib==1.5.3
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langsmith==0.7.6
openai==2.24.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
pandas==2.3.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.1
scikit-learn==1.7.2
PyYAML==6.0.3


